поправочки для Руслана :З
