+++ js/data.js
@@ -1,3 +1,5 @@
+import { getRandomNumber } from './util.js';
+
 const sms = [
   'Всё отлично!', 'В целом всё неплохо. Но не всё.',
   'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
@@ -12,4 +14,19 @@ const names = [
 
 //массив с именами
 
-export { sms, names };
+const createComment = () => ({
+  id: getRandomNumber(1, 1000000),
+  avatar: `img/avatar-${getRandomNumber(1, 6)}.svg`,
+  message: sms[getRandomNumber(0, 5)],
+  name: names[getRandomNumber(0, 6)]
+});
+
+const createObject = () => ({
+  id: getRandomNumber(1, 25),
+  url: `img/${getRandomNumber(1, 25)}.jpg`,
+  description: 'Строка, описание фотографии',
+  likes: getRandomNumber(15, 200),
+  comments: Array.from({ length: getRandomNumber(1, 100) }, createComment)
+});
+
+export { createObject };
