+++ js/main.js
@@ -1,4 +1,4 @@
 import { createObject } from './util.js';
-import './getPicture/js';
+import { } from './getPicture/js';
 
 const someObject = Array.from({ length: 25 }, createObject);
